fix: ajuste na rota de votação dos avaliadores
