+++ src/app/votacao/publica/confirmacao/avaliador/page.js
@@ -1,115 +1,65 @@
 "use client";
+import Button from "@/components/button";
 import ConfirmModal from "@/components/confirmModal";
-import React, { useState, useRef } from "react";
+import Image from "next/image";
+import { useRouter } from "next/navigation";
 
-const Header = () => (
-  <header className="w-full mb-4">
-    <h1 className="text-[#004854] text-2xl font-bold md:text-2xl">
-      AVALIAÇÃO
-    </h1>
-    <hr className="border-t-2 border-[#004854] my-3" />
-  </header>
-);
+export default function VotacaoPublica() {
+   const router = useRouter();
 
-const StarRating = ({ rating, setRating }) => (
-  <div className="flex flex-col gap-1">
-    <div className="flex gap-4">
-      {[...Array(5)].map((_, i) => (
-        <label key={i + 1}>
-          <input
-            type="radio"
-            value={i + 1}
-            className="hidden"
-            onClick={() => setRating(i + 1)}
-          />
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            className="bi bi-star-fill cursor-pointer w-10 h-10 md:w-12 md:h-12"
-            fill={i + 1 <= rating ? "#ffc107" : "#cccccc"}
-            viewBox="0 0 16 16"
-          >
-            <path d="M3.612 15.443c-.386.198-.824-.149-.746-.592l.83-4.73L.173 6.765c-.329-.314-.158-.888.283-.95l4.898-.696L7.538.792c.197-.39.73-.39.927 0l2.184 4.327 4.898.696c.441.062.612.636.282.95l-3.522 3.356.83 4.73c.078.443-.36.79-.746.592L8 13.187l-4.389 2.256z" />
-          </svg>
-        </label>
-      ))}
-    </div>
-  </div>
-);
-
-export default function RatingPage() {
-  const modalRef = useRef();
-  const [step, setStep] = useState(0);
-  const [acolhimento, setAcolhimento] = useState({ rating: null, comentario: "" });
-  const [inovacao, setInovacao] = useState({ rating: null, comentario: "" });
-  const [isCommentOpen, setIsCommentOpen] = useState(false);
-
-  const handleNextStep = () => {
-    if (step === 0) {
-      setStep(1);
-    } else {
-      const ratings = { acolhimento, inovacao };
-      console.log(ratings);
-      modalRef.current.openModal();
-    }
-  };
+   const handleRedirect = () => {
+     router.push("/votacao/publica/confirmacao/avaliador/classificacao");
+   };
 
   return (
-    <div className="flex flex-col px-6 py-6 items-center justify-center">
-      <Header />
-      <div className="flex flex-col w-full justify-end bg-[#F1F1F1] drop-shadow-md rounded-2xl gap-10 p-4 pl-8 pt-10 md:w-auto">
-        <span className="text-xl text-[#1A6C7C] font-bold">
-          {step === 0
-            ? "Como você avalia o nível de acolhimento deste projeto?"
-            : "Como você avalia o nível de inovação deste projeto?"}
-        </span>
-        <div className="flex w-full justify-center">
-          <StarRating
-            rating={step === 0 ? acolhimento.rating : inovacao.rating}
-            setRating={(value) =>
-              step === 0
-                ? setAcolhimento({ ...acolhimento, rating: value })
-                : setInovacao({ ...inovacao, rating: value })
-            }
-          />
-        </div>
-        <div className="flex flex-col gap-2">
-          <div className="flex justify-between">
-            <span className="text-xl text-[#1A6C7C]">Comentário</span>
-            <span className="text-[#1A6C7C] cursor-pointer select-none" onClick={() => setIsCommentOpen(!isCommentOpen)}>{isCommentOpen ? "Fechar" : "Opcional"}</span>
+    <main className="flex flex-col flex-grow items-center justify-start">
+      <div className="flex flex-col w-5/6 mt-4 md:mt-16">
+        <p className="font-sans font-regular text-left w-full text-[15px] md:text-[25px] text-[#004854] mb-[-8px] md:mb-[-16px]">
+          ESCOLHA DE
+        </p>
+        <p className="font-sans font-bold text-left w-full text-[23px] md:text-[36px] text-[#004854]">
+          PROJETO
+        </p>
+        <hr className="w-full border-[#004854] border-1 " />
+        <div className="flex flex-col md:flex-row md:w-5/6 gap-8 mt-6 md:gap-12 md:mt-16 md:mb-16 md:justify-stretch md:self-center">
+          <div className="order-2 w-170px w-4/5 md:w-1/3 h-fit bg-[#ffffff] rounded-b-[10px] flex flex-col shadow-gray-400 shadow-md self-center">
+            <div className="w-full aspect-video bg-[#cdcdcd] border-none p-0 place-content-center">
+              <Image
+                src="/undefinedImage.svg"
+                className="justify-self-center stroke-white w-[50px]"
+                alt="Imagem do Projeto"
+                width={50}
+                height={50}
+              />
+            </div>
+            <div className="flex flex-col px-5 py-3 hyphens-auto overflow-y-auto min-h-[100px] max-h-[150px] md:min-h-[150px] md:max-h-[200px]">
+              <p className="text-[#004854] font-regular text-[15px] md:text-[24px]">
+                {/* Nome do projeto */}
+              </p>
+              {/* <span className='text-[#004854] font-regular text-[24px]'>
+								</span> */}
+              <p className="text-[#004854] font-regular text-[10px] md:text-[18px]">
+                Descrição:
+                <span>{/* Descrição breve */}</span>
+              </p>
+            </div>
           </div>
-          <hr />
-          <textarea
-            placeholder="Digite o comentário"
-            className={isCommentOpen ? `w-full bg-white rounded-xl resize-none p-2` : `hidden`}
-            rows="4"
-            value={step === 0 ? acolhimento.comentario : inovacao.comentario}
-            onChange={(e) =>
-              step === 0
-                ? setAcolhimento({ ...acolhimento, comentario: e.target.value })
-                : setInovacao({ ...inovacao, comentario: e.target.value })
-            }
-          />
-        </div>
-        <div className="flex w-full justify-end gap-4">
-          {step > 0 ? (
-              <span
-                className="text-md text-[#1A6C7C] font-black cursor-pointer select-none"
-                onClick={() => setStep(step - 1)}
-              >
-                Voltar
-              </span>
-            ) : (
-              <span className="text-md text-[#1A6C7C] font-black cursor-pointer select-none">
-                Sair
-              </span>
-            )}
-          <span className="text-md text-[#1A6C7C] font-black cursor-pointer select-none" onClick={handleNextStep}>
-            {step === 0 ? "Próximo" : "Avaliar"}
-          </span>
+          <div className="order-1 md:order-2 md:w-2/3 flex flex-col content-between">
+            <div className="mb-auto">
+              <p className="text-[20px] font-bold mb-2 md:mb-4 md:text-[36px] text-[#004854]">
+                VOTO QUASE CONFIRMADO!
+              </p>
+              <p className="text-[15px] font-regular md:text-[25px] text-[#004854]">
+                Parabéns! Seu voto está quase confirmado, verifique abaixo se o
+                projeto é o desejado e confirme em seguida.
+              </p>
+            </div>
+            <div className="hidden md:flex flex-row space-x-6 mt-8 justify-center">
+              <Button onClick={handleRedirect} text={"VOTAR"} />
+            </div>
+          </div>
         </div>
       </div>
-
-      <ConfirmModal ref={modalRef} />
-    </div>
+    </main>
   );
 }
